requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.47.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
executing==2.2.0
flask==3.1.0
filelock==3.19.1
Flask==3.1.2
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
ollama==0.5.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
peft==0.17.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
trl==0.21.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
llama-cpp-python==0.2.90
